Adding UTexture2D to textures that can be put into a texture input node.
This was accomplished by changing the input type of the texture input node to "UTexture". This allows all objects derived from the UTexture class to be used, however only UTexture2D and UTextureRenderTarget2D are supported. If a user attempts to use a texture of other types, an error will be thrown.