Fix deadlock state when cooking synchronously
When synchronous cook is enabled, the GameThread would stall waiting for TouchEngine to process a frame.
During that processing, on a separate thread, one of the commands would schedule a task for the GameThread to recreate a texture, which would never happen, as the GameThread was stalled, effectivelly freezing Unreal.
Now we recreate the texture in a different thread, allowing the rest of the commands to execute, eventually unstalling the GameThread.