Only spawn separate thread if texture needs resize
On the render thread, we won't move to a separate thread just to come back to it later if we don't need to recreate the destination texture.

Had to move the final render thread code, which made the resource copy, to a function inside a struct in the render thread lambda, to be able to access private members of UTouchEngine. Otherwise it would have to be a member function and I didn't want to polute the header just because of this one-shot need. It couldn't be a lambda either, because the code that calls another render thread command can't capture a lambda in a lambda capture list.