Fix textures cleanup crash
Textures were getting cleaned up in the GameThread. That would conflict with operations in the RenderThread and cause DirectX to throw an exception.

Add a check to verify that UTouchEngine::CleanupTextures is called in the RenderThread and prevent the opposite from happening silently.

Add an early return to UTouchEngine::Clear in case it's already been cleaned up.